feat: #39 object 유틸함수 파일단위로 분리
