+++ src/utils/object/index.ts
@@ -1,0 +1,5 @@
+export { forIn } from './forIn';
+export { keys } from './keys';
+export { omit } from './omit';
+export { pick } from './pick';
+export { values } from './values';
