+++ src/utils/object/keys.ts
@@ -1,0 +1,3 @@
+export const keys = <T extends Record<string, unknown>>(object: T) => {
+  return Object.keys(object);
+};
