+++ src/utils/object/values.ts
@@ -1,0 +1,3 @@
+export const values = <T extends Record<string, unknown>>(object: T) => {
+  return Object.values(object);
+};
